New/requirements.txt
python-telegram-bot==21.6
requests
python-dotenv
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.32.3
